Simplify app
* Script simply calls app rather than creates
